Story Escape
Our game for the Game Engines 1 final project. To briefly explain, the goal is to grab a key located in a specific part of the map and reach the door at the end to complete the game. While doing this, you must navigate through obstacles such as saws, dragons, and avoid falling by completing challenging parkour sections.